2018/12/26
델타타임 적용
모든 단위가 프레임기준에서 시간기준으로 변경됨

+++ BEffect.h
@@ -32,7 +32,7 @@ public:
 	};
 protected:
 	//effType playing_EffectHandle;
-	int effect_Life;
+	float effect_Life;
 	
 	ML::Vec3 pos;//エフェクト再生位置
 	ML::Vec3 target;//ターゲット位置(ある場合だけもらう)
